refactor: fetch함수 추상화
