+++ src/hooks/useFetchCart.ts
@@ -1,41 +1,15 @@
-import { useRecoilValue, useSetRecoilState } from 'recoil';
-import { baseURLSelector } from '../store/server';
-import { AUTH } from '../constants/auth';
-import { useCallback } from 'react';
+import { useSetRecoilState } from 'recoil';
 import { cartAtom } from '../store/cart';
 import { Cart } from '../types/responseData';
-import { FetchMethod } from '../types/global';
+import useFetch from './useFetch';
 
 const useFetchCart = () => {
-  const baseURL = useRecoilValue(baseURLSelector);
   const setCartList = useSetRecoilState(cartAtom);
-
-  const handleCartItems = useCallback(
-    async (method: FetchMethod, body: {}, id?: number) => {
-      const response = await fetch(
-        `${baseURL}/cart-items${id ? `/${id}` : ''}`,
-        {
-          method,
-          body: JSON.stringify(body),
-          headers: {
-            'Content-Type': 'application/json',
-            authorization: `Basic ${AUTH}`,
-          },
-        }
-      );
-
-      if (!response.ok) throw new Error(`error code : ${response.status}`);
-
-      const data = await response.text();
-      if (!data) return null;
-      return JSON.parse(data);
-    },
-    [baseURL]
-  );
+  const { handleFetch } = useFetch('/cart-items');
 
   const addToCart = async (productId: number) => {
     try {
-      const data = await handleCartItems('POST', { productId });
+      const data = await handleFetch('POST', { productId });
       const { id, quantity, product } = data;
 
       setCartList((cartList) => [...cartList, { id, quantity, product }]);
@@ -46,7 +20,7 @@ const useFetchCart = () => {
 
   const updateCartItem = async (id: number, quantity: number) => {
     try {
-      handleCartItems('PATCH', { quantity }, id);
+      handleFetch('PATCH', { quantity }, id);
       setCartList(
         (cartList) =>
           [
@@ -63,7 +37,7 @@ const useFetchCart = () => {
 
   const deleteCartItem = async (id: number) => {
     try {
-      handleCartItems('DELETE', {}, id);
+      handleFetch('DELETE', {}, id);
       setCartList((cartList) => [...cartList.filter((item) => item.id !== id)]);
     } catch (error) {
       alert(error);
